+++ src/clindex/utils.clj
@@ -120,12 +120,13 @@
   (stable-id :spec-alpha namespace-symb spec-key))
 
 (defn rectangle-select
-  "Given a string with lines, returns a string with the subregion
+  "Given a vector of strings, assumes each string is a line and
+  returns a string with the subregion
   between l1 and l2 lines and c1 and c2 columns.
   No c2 means to the end of the line."
-  ([s l1 l2 c1] (rectangle-select s l1 l2 c1 nil))
-  ([s l1 l2 c1 c2]
-   (->> (str/split-lines s)
+  ([lines l1 l2 c1] (rectangle-select lines l1 l2 c1 nil))
+  ([lines l1 l2 c1 c2]
+   (->> lines
         (drop (dec l1))
         (take (inc (- l2 l1)))
         (map (fn [l]